Wolf : Réparation du wolf. Le jeu se ferme si le PLAYER_START est dans un mur. Flag Oa remis. Le shading est maintenant en double et rend en fondu.
Build : 0.09

+++ srcs/world.c
@@ -34,7 +34,7 @@ static void	determine_colors(t_ray r, int *color)
 static int	shading(t_data *data, t_ray r, int color)
 {
 	int c;
-	int	i;
+	double	i;
 
 	if (data->map.t == DAY)
 	{
